chore(core): separate values module
The values() function started as a tiny helper which just returned the values property of the data object, but has now grown into a much more complicated thing which performs remote fetching, orchestrates caching, and transforms data sets. It deserves a dedicated module, which leaves the data.js module more clearly focused on doing the preprocessing required for different chart forms.
