+++ source/feature.js
@@ -2,7 +2,7 @@ import { encodingType, encodingValue } from './encodings.js'
 import { layerTestRecursive } from './views.js'
 import { mark } from './helpers.js'
 import { memoize } from './memoize.js'
-import { values } from './data.js'
+import { values } from './values.js'
 
 // this is a mistake
 // import { isPresent } from '@ember/utils';
